postgres_to_es/requirements.txt
psycopg2-binary==2.8.6
requests==2.24.0
backoff==1.10.0
pydantic==1.6.1
python-decouple==3.3
flake8==3.8.4